refactor|chore(CPP-104): editorial version upgrade
