+++ src/pages/404.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { navigate } from '@reach/router';
 
 export default () => {
-  useEffect(() => {
-    navigate('/', { replace: true });
-  }, []);
-  return null;
+	useEffect(() => {
+		navigate('/', { replace: true });
+	}, []);
+	return null;
 };
